refactor(event): keep deprecated ReplyEndReason in event, privatize classify_error
- Move the deprecated ReplyEndReason back to event/_event.py (original
  3-member impl) with a @deprecated marker, instead of duplicating it in
  types. types now owns only ReplyFinishedReason / ErrorType / ErrorInfo.
- agentscope.event.ReplyEndReason still works (re-exported from _event);
  agentscope.types no longer exports it.
- Rename classify_error -> _classify_error to mark it internal.
